+++ lib/fastlane/plugin/wpmreleasetoolkit/bumpers/version_bumper.rb
@@ -1,4 +1,4 @@
-require_relative '../models/version'
+require_relative '../models/app_version'
 
 module Fastlane
   module Bumper
