+++ lib/fastlane/plugin/wpmreleasetoolkit/formatters/ios_version_formatter.rb
@@ -1,4 +1,5 @@
-require_relative '../models/version'
+require_relative '../models/app_version'
+require_relative 'version_formatter'
 
 module Fastlane
   module Formatter
