+++ lib/fastlane/plugin/wpmreleasetoolkit/formatters/version_formatter.rb
@@ -1,4 +1,4 @@
-require_relative '../models/version'
+require_relative '../models/app_version'
 
 module Fastlane
   module Formatter
